Fix a compile-time warning on amd64, for the second argument to the getline() call - incompatible pointer type.
I'm assuming that size_t is equivalent to uint32_t
on a 32-bit architecture, but not on a 64-bit arch.

A quick test showed no difference in actual output.

+++ utils/exchange2mbox.c
@@ -120,7 +120,7 @@ static uint32_t update(TALLOC_CTX *mem_ctx, FILE *fp,
 {
 	enum MAPISTATUS		retval;
 	struct mapi_profile	profile;
-	uint32_t		read_size;
+	size_t			read_size;
 	char			*line = NULL;
 	size_t			size;
 	const char		*msgid;
